structure refactor of BinaryTreeConstructors + FastIntersection default
Removed all the unnecessary material
